Added AppUser class and UniversityEmailDatabase
This commit includes the class of the user utilised by our app, the name was changed to AppUser instead of FirebaseUserAdapter since we now have a specific firebase login service to interact with the backend, the old name would be misleading. A UniversityEmailDatabse class was added, this contains a set of trusted email providers and offers a public method **has(String email)** (used in AppUser **hasUniversityEmail()** method). This was implemented in a separate file since it will be useful for other classes and it would be unnecessary for each instance of user to store this database. Tests for this part are still not finished, a AppUserTest.java JUnit file will be uploaded soon.
